sanitizer: fix three ASan/UBSan failures
- z_ascache.cpp: rename local CV_OnOff[] to cv_async_onoff_values[] to
  avoid ASAN ODR collision with the global CV_OnOff in command.cpp
- m_fixed.h: cast through uint32_t before left-shifting to suppress
  UBSan signed-integer-overflow on negative fixed_t construction
- nodebuild_utility.cpp: guard &Vertices[0] with a size check in both
  SelectVertexExact and SelectVertexClose to prevent UBSan null-pointer
  binding when the TArray backing pointer is null

+++ src/include/core/m_fixed.h
@@ -66,7 +66,7 @@ class fixed_t
     }
     inline fixed_t(int a)
     {
-        val = a << FBITS;
+        val = (int32_t)((uint32_t)a << FBITS);
     }
     /// Construct from a raw 16.16 value that has already been shifted (no additional shift applied)
     inline fixed_t(int a, bool rawShifted)
